bugfix: locale was not changing based on the users session
By using a DEFAULT_OPTIONS constant, we were setting the locale to the locale of the first request (usually en).

To get around this, make it a class method which we call when creating the configurations.

Also added checking of the locale. If it doesn't exist in TinyMCE, default to English.